The Utf8FrameValidator only handles its own protocol violation exception (#12210)
Motivation:

Currently `Utf8FrameValidator` catches all `CorruptedWebSocketFrameException` and trying to close the channel if it's open. This is incorrect behavior because `WebsocketFrameDecoder` throws the same type of exception, but with a different status code and reason, and does not close the channel if the `closeOnProtocolViolation` property is set to false.

Modification:

Correct that the Utf8FrameValidator only handles its own protocol violation exception  and sends the CloseWebsocketFrame with the appropriate status code and reason. Pass the `closeOnProtocolViolation` property to the `Utf8FrameValidator`;

Result:
The Utf8FrameValidator only handle its own protocol violation exception and consolidate with decoder config.