Update to the latest enforcer plug-in (#12363)
Motivation:
Since we baseline on Java 11, we need a 3.x version of the enforcer plug-in anyway.
This is no longer a special case.

Modification:
Update the required enforcer plug-in version so we don't have to create a special case for it, for every Java release.

Result:
Build now works on Java 18, and likely future Java versions as well.