Remove null check (#12367)
Motivation:
There is a `null` check for the `Process` instance. However, it can never be `null` because it is being initialized at the first line and if there is an exception during initialization then we will throw the exception directly.

Modification:
Removed useless `null` check

Result:
No need for a `null` check and simpler code.